created ConversationBody responsible for rendering conversation messages sent by users
