+++ client/src/app/(main)/explore/[id]/page.tsx
@@ -1,9 +1,9 @@
-import { ConversationBody } from "@/ui/messaging/ConversationBody";
+import { Conversation } from "@/ui/messaging/Conversation";
 export default function Page({ params }: { params: { id: string } }) {
   const { id: conversationId } = params;
   return (
     <div>
-      <ConversationBody conversationId={conversationId} />
+      <Conversation conversationId={conversationId} />
     </div>
   );
 }
